+++ app/src/main/java/com/example/partymaker/ui/security/SecurityScanActivity.java
@@ -11,8 +11,8 @@ import android.widget.Toast;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.content.FileProvider;
 import com.example.partymaker.R;
-import com.example.partymaker.utils.security.SecurityAgent;
-import com.example.partymaker.utils.security.SecurityReport;
+import com.example.partymaker.utils.security.monitoring.SecurityAgent;
+import com.example.partymaker.utils.security.monitoring.SecurityReport;
 import java.io.File;
 import java.io.FileWriter;
 import java.text.SimpleDateFormat;
